+++ NEWS.rst
@@ -19,6 +19,8 @@ Version history
 - Rebuild the Snap package to include updated system packages.
   See https://ubuntu.com/security/notices/USN-5495-1/.
 
+- Add new CLI option to provide secrets via a YAML-based configuration file.
+
 0.0.17 (2022-06-17)
 ===================